refactor to more closely match code style and prep for webpack patch in preload
