+++ src/app/main.ts
@@ -1,5 +1,5 @@
 import { members } from "./core/badges-flairs";
-import webpackPush from "./webpackInject";
+//import webpackPush from "./webpackInject";
 import ReGuilded from "./core/ReGuilded";
 //import { ipcRenderer } from "electron";
 
@@ -10,24 +10,24 @@ window.ReGuilded = new ReGuilded();
 //     require(preload);
 // }
 
-function setPush(obj) {
-    Object.defineProperty(window.webpackJsonp, "push", obj)
-}
+// function setPush(obj) {
+//     Object.defineProperty(window.webpackJsonp, "push", obj)
+// }
 
-document.addEventListener("readystatechange", () => {
-    // To wait for the bundle to be created
-    if (document.readyState === "interactive")
-        // Wait when bundle loads
-        window.bundle.addEventListener("load", () => {
-            // Saves the old push
-            window.webpackJsonp._push = window.webpackJsonp.push;
+// document.addEventListener("readystatechange", () => {
+//     // To wait for the bundle to be created
+//     if (document.readyState === "interactive")
+//         // Wait when bundle loads
+//         window.bundle.addEventListener("load", () => {
+//             // Saves the old push
+//             window.webpackJsonp._push = window.webpackJsonp.push;
 
-            setPush({
-                get: () => webpackPush.bind(window.webpackJsonp),
-                set: (value) => setPush({get: () => value})
-            });
-        });
-});
+//             setPush({
+//                 get: () => webpackPush.bind(window.webpackJsonp),
+//                 set: (value) => setPush({get: () => value})
+//             });
+//         });
+// });
 // Fetch ReGuilded things
 (async () => {
     // Global badge holders
