+++ src/preload/main.ts
@@ -4,11 +4,16 @@ import { contextBridge, ipcRenderer, shell } from "electron";
 import getSettingsFile from "./get-settings";
 import SettingsManager from "./settings";
 import { join } from "path";
+import webpackPush from "./webpackInject";
 
 const settingsPath = join(__dirname, "./settings");
 
 (async() => {
     const reGuildedConfigAndSettings = async () => {
+        function setPush(obj) {
+            Object.defineProperty(window.webpackJsonp, "push", obj)
+        }
+
         const settingsManager = new SettingsManager(
             join(__dirname, "./settings"),
             await getSettingsFile(settingsPath)
@@ -41,12 +46,24 @@ const settingsPath = join(__dirname, "./settings");
                 shell.openExternal(path);
             }
         });
+        document.addEventListener("readystatechange", () => {
+            // To wait for the bundle to be created
+            if (document.readyState === "interactive")
+                // Wait when bundle loads
+                window.bundle.addEventListener("load", () => {
+                    // Saves the old push
+                    window.webpackJsonp._push = window.webpackJsonp.push;
+        
+                    setPush({
+                        get: () => webpackPush.bind(window.webpackJsonp),
+                        set: (value) => setPush({get: () => value})
+                    });
+                });
+        });
     };
     await reGuildedConfigAndSettings()
         .then(() => {
             const preload = ipcRenderer.sendSync("REGUILDED_GET_PRELOAD");
-            if(preload) {
-                import(preload);
-            };
+            if(preload) import(preload);
         }).catch(console.error);
 })();
